+++ examples/hono-official-examples/public/custom-event.js
@@ -1,0 +1,11 @@
+const target = document.getElementById("custom-event-target")
+
+setInterval(() => {
+  target?.dispatchEvent(
+    new CustomEvent("myevent", {
+      detail: JSON.stringify({
+        eventTime: new Date().toLocaleTimeString()
+      })
+    })
+  )
+}, 1000)